backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
PyMuPDF
python-docx
opencv-python
faker
spacy
pytesseract